Refactoring and code cleanup
- rename CommandParser to StatementParser
- move tests from test_shlexparsing.py to test_parsing.py
- standardize the output of the parse() method into a variable called statement.
